requirements.txt
# Using vowpal-wabbit-next instead of standard vowpalwabbit
# Reason: vowpal-wabbit-next provides pre-built wheels and doesn't require C++ compilation
# The standard vowpalwabbit requires cmake, C++ compiler, and system dependencies
vowpal-wabbit-next>=0.7.0
pyyaml>=6.0
pyarrow>=14.0.0

# Data generation with LLM providers (optional)
anthropic>=0.18.0        # For Claude Haiku 4.5
google-generativeai>=0.3.0  # For Gemini 2.0 Flash
openai>=1.0.0            # For GPT-4o-mini
tqdm>=4.65.0             # Progress bars

# Embedding-based classification (optional, for better accuracy)
sentence-transformers>=2.0.0  # For semantic embeddings
scikit-learn>=1.3.0           # For fast classifiers
xgboost>=2.0.0                # For gradient boosting (optional)
lightgbm>=4.0.0               # For fast gradient boosting (optional)
